pipeline/requirements.txt
aws-cdk-lib==2.51.1
constructs>=10.0.0,<11.0.0
aws-lambda-powertools==2.3.1
aws-cdk.aws-lambda-python-alpha==2.51.1a0